+++ src/Options.php
@@ -4,7 +4,12 @@ declare(strict_types=1);
 
 namespace Sentry;
 
+use Sentry\Integration\ErrorListenerIntegration;
+use Sentry\Integration\ExceptionListenerIntegration;
+use Sentry\Integration\FatalErrorListenerIntegration;
 use Sentry\Integration\IntegrationInterface;
+use Sentry\Integration\RequestIntegration;
+use Sentry\Integration\TransactionIntegration;
 use Symfony\Component\OptionsResolver\Options as SymfonyOptions;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -49,6 +54,11 @@ final class Options
      * @var OptionsResolver The options resolver
      */
     private $resolver;
+
+    /**
+     * @var IntegrationInterface[]|null The list of default integrations
+     */
+    private $defaultIntegrations;
 
     /**
      * Class constructor.
@@ -504,11 +514,13 @@ final class Options
     }
 
     /**
-     * Set integrations that will be used by the created client.
+     * Sets the list of integrations that should be installed after SDK was
+     * initialized or a function that receives default integrations and returns
+     * a new, updated list.
      *
-     * @param IntegrationInterface[] $integrations The integrations
+     * @param IntegrationInterface[]|callable $integrations The list or callable
      */
-    public function setIntegrations(array $integrations): void
+    public function setIntegrations($integrations): void
     {
         $options = array_merge($this->options, ['integrations' => $integrations]);
 
@@ -522,7 +534,23 @@ final class Options
      */
     public function getIntegrations(): array
     {
-        return $this->options['integrations'];
+        $defaultIntegrations = $this->getDefaultIntegrations();
+        $userIntegrations = $this->options['integrations'];
+        $integrations = [];
+
+        if (\is_callable($userIntegrations)) {
+            return $userIntegrations($defaultIntegrations);
+        }
+
+        foreach ($defaultIntegrations as $defaultIntegration) {
+            $integrations[\get_class($defaultIntegration)] = $defaultIntegration;
+        }
+
+        foreach ($userIntegrations as $userIntegration) {
+            $integrations[\get_class($userIntegration)] = $userIntegration;
+        }
+
+        return array_values($integrations);
     }
 
     /**
@@ -751,7 +779,7 @@ final class Options
         $resolver->setAllowedTypes('error_types', ['int']);
         $resolver->setAllowedTypes('max_breadcrumbs', 'int');
         $resolver->setAllowedTypes('before_breadcrumb', ['callable']);
-        $resolver->setAllowedTypes('integrations', 'array');
+        $resolver->setAllowedTypes('integrations', ['array', 'callable']);
         $resolver->setAllowedTypes('send_default_pii', 'bool');
         $resolver->setAllowedTypes('default_integrations', 'bool');
         $resolver->setAllowedTypes('max_value_length', 'int');
@@ -903,10 +931,14 @@ final class Options
      * Validates that the elements of this option are all class instances that
      * implements the {@see IntegrationInterface} interface.
      *
-     * @param array $integrations The value to validate
+     * @param array|callable $integrations The value to validate
      */
-    private function validateIntegrationsOption(array $integrations): bool
+    private function validateIntegrationsOption($integrations): bool
     {
+        if (\is_callable($integrations)) {
+            return true;
+        }
+
         foreach ($integrations as $integration) {
             if (!$integration instanceof IntegrationInterface) {
                 return false;
@@ -956,5 +988,29 @@ final class Options
         }
 
         return true;
+    }
+
+    /**
+     * Gets the list of default integrations.
+     *
+     * @return IntegrationInterface[]
+     */
+    private function getDefaultIntegrations(): array
+    {
+        if (!$this->options['default_integrations']) {
+            return [];
+        }
+
+        if (null === $this->defaultIntegrations) {
+            $this->defaultIntegrations = [
+                new ExceptionListenerIntegration(),
+                new ErrorListenerIntegration(null, false),
+                new FatalErrorListenerIntegration(),
+                new RequestIntegration(),
+                new TransactionIntegration(),
+            ];
+        }
+
+        return $this->defaultIntegrations;
     }
 }