+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Update PHPStan and introduce Psalm (#846)
 - Add an integration to set the transaction attribute of the event (#865)
 - Deprecate `Hub::getCurrent` and `Hub::setCurrent` methods to set the current hub instance (#847)
+- Make the `integrations` option accept a `callable` that will receive the list of default integrations and returns a customized list (#919)
 
 ## 2.1.3 (2019-09-06)